+++ src/main/java/net/minecraft/client/renderer/vertex/VertexFormat.java
@@ -50,7 +50,7 @@ public class VertexFormat {
             return this;
         } else {
             this.elements.add(element);
-            this.offsets.add(Integer.valueOf(this.nextOffset));
+            this.offsets.add(this.nextOffset);
 
             switch (element.getUsage()) {
                 case NORMAL:
@@ -62,7 +62,7 @@ public class VertexFormat {
                     break;
 
                 case UV:
-                    this.uvOffsetsById.add(element.getIndex(), Integer.valueOf(this.nextOffset));
+                    this.uvOffsetsById.add(element.getIndex(), this.nextOffset);
             }
 
             this.nextOffset += element.getSize();
@@ -91,7 +91,7 @@ public class VertexFormat {
     }
 
     public int getUvOffsetById(int id) {
-        return this.uvOffsetsById.get(id).intValue();
+        return this.uvOffsetsById.get(id);
     }
 
     public String toString() {
@@ -143,7 +143,7 @@ public class VertexFormat {
     }
 
     public int getOffset(int p_181720_1_) {
-        return this.offsets.get(p_181720_1_).intValue();
+        return this.offsets.get(p_181720_1_);
     }
 
     public boolean equals(Object p_equals_1_) {
