+++ src/main/java/net/minecraft/entity/passive/EntityPig.java
@@ -48,7 +48,7 @@ public class EntityPig extends EntityAnimal {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, Byte.valueOf((byte) 0));
+        this.dataWatcher.addObject(16, (byte) 0);
     }
 
     public void writeEntityToNBT(NBTTagCompound tagCompound) {
@@ -114,9 +114,9 @@ public class EntityPig extends EntityAnimal {
 
     public void setSaddled(boolean saddled) {
         if (saddled) {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte) 1));
+            this.dataWatcher.updateObject(16, (byte) 1);
         } else {
-            this.dataWatcher.updateObject(16, Byte.valueOf((byte) 0));
+            this.dataWatcher.updateObject(16, (byte) 0);
         }
     }
